feature/BRS-add-utin-test-for-service

+++ internal/repository/postgres_db/credit/interfaces_repository.go
@@ -12,6 +12,6 @@ type ICreditRepository interface {
 	GetByUser(ctx context.Context, idUser int64) ([]core.Credit, error)
 	GetById(ctx context.Context, id int64) (*core.Credit, error)
 	Create(ctx context.Context, input *core.Credit) (*core.Credit, error)
-	Repay(ctx context.Context, id int64, amount int) (*core.Credit, error)
+	Repay(ctx context.Context, id int64, amount int64) (*core.Credit, error)
 	Delete(ctx context.Context, id int64) (int64, error)
 }
